+++ changelog.md
@@ -21,6 +21,10 @@
   * A view that uses `bindTo` for its own "close" event will have it's close handler called correctly
   * Returning `false` from the `onBeforeClose` method will prevent the view from being closed
 
+* Module
+  * Triggers a "before:stop" event/method before the module is stopped
+  * Triggers a "stop" event/method after the module has been stopped
+
 ### v1.0.0-beta6 [view commit logs](https://github.com/marionettejs/backbone.marionette/compare/v1.0.0-beta5...v1.0.0-beta6)
 
 * CollectionView / CompositeView